revise

+++ Competitive_Programming/Computation_Basic/divisor.hpp
@@ -4,7 +4,7 @@ template<typename T>
 vector<T> divisor(T n)
 {
     vector<T> res;
-    for(T i=1;(ll)i*i<=n;i++){
+    for(T i=1;(long long)i*i<=n;i++){
         if(n%i==0){
             res.push_back(i);
             if(i != n/i){
